+++ src/main/java/mediathek/filmlisten/FilmeLaden.java
@@ -9,6 +9,7 @@ import mediathek.daten.ListeFilme;
 import mediathek.filmeSuchen.ListenerFilmeLaden;
 import mediathek.filmeSuchen.ListenerFilmeLadenEvent;
 import mediathek.filmlisten.reader.FilmListReader;
+import mediathek.gui.duplicates.CommonStatsEvaluationTask;
 import mediathek.gui.duplicates.FilmDuplicateEvaluationTask;
 import mediathek.gui.messages.FilmListReadStopEvent;
 import mediathek.gui.tasks.BlacklistFilterWorker;
@@ -317,8 +318,13 @@ public class FilmeLaden {
             throw new RuntimeException(e);
         }
         var workerTask = CompletableFuture.runAsync(new RefreshAboWorker(progLabel, progressBar))
-                .thenRun(new BlacklistFilterWorker(progLabel, progressBar))
-                .thenRun(new FilmDuplicateEvaluationTask());
+                .thenRun(new BlacklistFilterWorker(progLabel, progressBar));
+
+        var evaluateDuplicates = ApplicationConfiguration.getConfiguration().getBoolean(ApplicationConfiguration.FILM_EVALUATE_DUPLICATES, true);
+        if (evaluateDuplicates) {
+            workerTask = workerTask.thenRun(new FilmDuplicateEvaluationTask());
+        }
+        workerTask = workerTask.thenRun(new CommonStatsEvaluationTask());
 
         if (writeFilmList) {
             workerTask = workerTask.thenRun(new FilmlistWriterWorker(progLabel, progressBar));